+++ generators/app/npm.js
@@ -1,0 +1,75 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file npm on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 18/11/2017
+ * Time: 16:06
+ * Version: 1.0.0
+ */
+(function () {
+	'use strict';
+
+	const npmModules = [
+		'app',
+		'autoprefixer',
+		'autoprefixer-core',
+		'browser-window',
+		'cssnano',
+		'graceful-fs',
+		'grunt',
+		'grunt-angular-file-loader',
+		'grunt-angular-templates',
+		'grunt-autoprefixer',
+		'grunt-cache-pug-compile',
+		'grunt-concurrent',
+		'grunt-contrib-clean',
+		'grunt-contrib-concat',
+		'grunt-contrib-copy',
+		'grunt-contrib-cssmin',
+		'grunt-contrib-htmlmin',
+		'grunt-contrib-imagemin',
+		'grunt-contrib-jshint',
+		'grunt-contrib-less',
+		'grunt-contrib-pug',
+		'grunt-contrib-uglify',
+		'grunt-cssnano',
+		'grunt-filerev',
+		'grunt-google-cdn',
+		'grunt-jscs',
+		'grunt-less-imports',
+		'grunt-merge-json',
+		'grunt-newer',
+		'grunt-notify',
+		'grunt-pixrem',
+		'grunt-postcss',
+		'grunt-preprocess',
+		'grunt-prettier',
+		'grunt-string-replace',
+		'grunt-svgmin',
+		'grunt-usemin',
+		'grunt-wiredep',
+		'jasmine-core',
+		'jit-grunt',
+		'jshint-stylish',
+		'karma',
+		'karma-jasmine',
+		'karma-phantomjs-launcher',
+		'less-plugin-autoprefix',
+		'minimatch',
+		'ng-simple-webrtc',
+		'phantomjs-prebuilt',
+		'pixrem',
+		'prettier',
+		'pug',
+		'pug-inheritance',
+		'time-grunt'
+	];
+
+	module.exports = class Npm {
+		static getModules() {
+			return npmModules;
+		}
+	};
+
+})();