+++ CHANGELOG.md
@@ -6,6 +6,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.1.2]
+### Added
+- New badge for MIT License
+- New badge for CodeClimate maintainability
+- New badge for CodeClimate test coverage
+
+### Changed
+- The method install in index.js now call npm and bower private methods
+- The colors are now imported from colors module
+- The npm modules are now imported from npm module
+- The bower modules are now imported from bower module
+
 ## [0.1.1]
 ### Changed
 - coveralls badge